feat: define 'exact' scheme for sui
Define the initial specification to be used for the 'exact' scheme for
the Sui Blockchain.

This specification is designed to be optimially flexible for clients to
be able to fully leverage PTBs (Programmable Transaction Blocks) to
either provide a direct payment or perform a set of operations to swap
from one asset to the asset expected by the resource server. It also
provides the optionality for leveraging sponsorship/gas-less
transactions.