feat: Add x402-browser-fetch
This commit includes a new package for x402/typescript called
`x402-browser-fetch`. The existing `x402-fetch` package is not
browser/client compatible due to how the account and signing data
function is passed. `x402-browser-fetch` accepts a
`SignTypedDataFunction` parameter which allows for any kind of wallet or
account implementation to be used with x402 server requests