After a long fought battle with the jwt implementation, we have a hack of a solution by ripping the rust implementation from the cdp-sdk repo and modifying it to suit our needs until we have a better solution.
In addition to that, we have an MVP structure passed to coinbase sdk facilitator. We need to move it into its own module as a drop-in facilitator.

We aren't getting an error yet, since we are not passing a valid value in, but it is a good stepping stone towards enlightenment.